texts and channel existence

+++ eventfunctions/gags/chocolate.js
@@ -2,6 +2,7 @@ const { getOption } = require("../../functions/configfunctions")
 const { messageSendChannel } = require("../../functions/messagefunctions")
 const { getUserVar, setUserVar } = require("../../functions/usercontext")
 const { getGag, assignGag, deleteGag} = require("../../functions/gagfunctions.js");
+const { getPronouns } = require("../../functions/pronounfunctions.js");
 
 const DISSOLVE_RATE_MS = 60000;
 
@@ -12,19 +13,19 @@ async function functiontick(userID) {
     }
 
     // Decrement Intensity every timer interval
-    if (getUserVar(userID, "confectionaryDissolveTimer") < Date.now() && getGag(userID, "chocolate")) {
+    if (getUserVar(userID, "confectionaryDissolveTimer") < Date.now() && getGag(userID, "chocolate") && process.recentmessages[userID]) {
         if(getGag(userID, "chocolate").intensity > 1){
             setUserVar(userID, "confectionaryDissolveTimer", Date.now() + DISSOLVE_RATE_MS)
             // Get Intensity and push decremented version
             let oldIntensity = getGag(userID, "chocolate").intensity
             assignGag(userID, "chocolate", oldIntensity - 1)
-            messageSendChannel(`<@${userID}>'s chocolate gag has shrunk~`, process.recentmessages[userID])
+            messageSendChannel(`<@${userID}>'s licking has shrunk ${getPronouns(userID, "reflexive")} Chocolate Gag a little bit!`, process.recentmessages[userID])
         }
         else {
             // Clear Gag and Dissolve Timer
             setUserVar(userID, "confectionaryDissolveTimer", undefined)
             deleteGag(userID, "chocolate")
-            messageSendChannel(`<@${userID}>'s chocolate gag has dissolved away~`, process.recentmessages[userID])
+            messageSendChannel(`<@${userID}>'s Chocolate Gag has dissolved away!`, process.recentmessages[userID])
         }
     }
 }
