+++ eventfunctions/gags/gummy.js
@@ -12,19 +12,19 @@ async function functiontick(userID) {
     }
 
     // Decrement Intensity every timer interval
-    if (getUserVar(userID, "confectionaryDissolveTimer") < Date.now() && getGag(userID, "gummy")) {
+    if (getUserVar(userID, "confectionaryDissolveTimer") < Date.now() && getGag(userID, "gummy") && process.recentmessages[userID]) {
         if(getGag(userID, "gummy").intensity > 1){
             setUserVar(userID, "confectionaryDissolveTimer", Date.now() + DISSOLVE_RATE_MS)
             // Get Intensity and push decremented version
             let oldIntensity = getGag(userID, "gummy").intensity
             assignGag(userID, "gummy", oldIntensity - 1)
-            messageSendChannel(`<@${userID}>'s gummy gag has shrunk~`, process.recentmessages[userID])
+            messageSendChannel(`<@${userID}>'s licking has shrunk ${getPronouns(userID, "reflexive")} Gummy Gag a little bit!`, process.recentmessages[userID])
         }
         else {
             // Clear Gag and Dissolve Timer
             setUserVar(userID, "confectionaryDissolveTimer", undefined)
             deleteGag(userID, "gummy")
-            messageSendChannel(`<@${userID}>'s gummy gag has dissolved away~`, process.recentmessages[userID])
+            messageSendChannel(`<@${userID}>'s Gummy Gag has dissolved away!`, process.recentmessages[userID])
         }
     }
 }
