+++ eventfunctions/gags/jawbreaker.js
@@ -12,19 +12,19 @@ async function functiontick(userID) {
     }
 
     // Decrement Intensity every timer interval
-    if (getUserVar(userID, "confectionaryDissolveTimer") < Date.now() && getGag(userID, "jawbreaker")) {
+    if (getUserVar(userID, "confectionaryDissolveTimer") < Date.now() && getGag(userID, "jawbreaker") && process.recentmessages[userID]) {
         if(getGag(userID, "jawbreaker").intensity > 1){
             setUserVar(userID, "confectionaryDissolveTimer", Date.now() + DISSOLVE_RATE_MS)
             // Get Intensity and push decremented version
             let oldIntensity = getGag(userID, "jawbreaker").intensity
             assignGag(userID, "jawbreaker", oldIntensity - 1)
-            messageSendChannel(`<@${userID}>'s jawbreaker gag has shrunk~`, process.recentmessages[userID])
+            messageSendChannel(`<@${userID}>'s licking has shrunk ${getPronouns(userID, "reflexive")} Jawbreaker Gag a little bit!`, process.recentmessages[userID])
         }
         else {
             // Clear Gag and Dissolve Timer
             setUserVar(userID, "confectionaryDissolveTimer", undefined)
             deleteGag(userID, "jawbreaker")
-            messageSendChannel(`<@${userID}>'s jawbreaker gag has dissolved away~`, process.recentmessages[userID])
+            messageSendChannel(`<@${userID}>'s Jawbreaker Gag has dissolved away!`, process.recentmessages[userID])
         }
     }
 }
